torch/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.0
einops>=0.7.0
tqdm>=4.65.0
wandb>=0.16.0
pyyaml>=6.0
numpy>=1.24.0
Pillow>=9.0.0
scipy>=1.11.0
